Napravljen delete button test
