+++ tv_shows_app/src/components/features/reviews/ReviewItem/ReviewItem.test.tsx
@@ -5,21 +5,7 @@ import { IReview, ReviewItem } from './ReviewItem';
  * 2) contain correct rating
  * 3) contain correct review comment
  * 4) render delete button
- * 5) call onDelete callback only once
- * 6) open modal on click and call removeReview and mutate on confirm
  */
-
-jest.mock('@/fetchers/mutators', () => {
-  return {
-    deleteReviewMutator: jest.fn().mockResolvedValue(null),
-  };
-});
-
-jest.mock('swr', () => {
-  return {
-    mutate: jest.fn(),
-  };
-});
 
 describe('ReviewItem', () => {
   const mockReview: IReview = {
@@ -27,7 +13,7 @@ describe('ReviewItem', () => {
     rating: 2,
     show_id: 2,
     user: {
-      id: 0,
+      id: -1,
       email: 'test@test.com',
       image_url: '',
     },
@@ -55,22 +41,12 @@ describe('ReviewItem', () => {
     const reviewCommentDisplay = screen.getByText(mockReview.comment);
     expect(reviewCommentDisplay).toBeInTheDocument();
   });
-  /** 
 
   it('should render delete button', () => {
-    render(<ReviewItem review={mockReview} onRemove={() => {}} />);
-    const deleteButton = screen.getByRole('button');
+    render(
+      <ReviewItem review={mockReview} onRemove={() => {}} onEdit={() => {}} />
+    );
+    const deleteButton = screen.getByTestId('delete-button');
     expect(deleteButton).toBeInTheDocument();
   });
-  it('should call onDelete callback only once', () => {
-    const mockOnDelete = jest.fn();
-    render(<ReviewItem review={mockReview} onRemove={mockOnDelete} />);
-    const deleteButton = screen.getByRole('button');
-    deleteButton.click();
-    expect(mockOnDelete).toHaveBeenCalled();
-    expect(mockOnDelete).toHaveBeenCalledTimes(1);
-    expect(mockOnDelete).toHaveBeenCalledWith(mockReview);
-  });
-  */
-  it('should open modal on click and call removeReview and mutate on confirm', () => {});
 });
